chore: extract-module arrays/cow
