+++ src/arrays/mod.rs
@@ -1,3 +1,5 @@
+mod cow;
+
 use std::fmt;
 use std::iter;
 
@@ -12,6 +14,8 @@ use num::complex::Complex64;
 use num::{BigInt, BigRational, Zero};
 use num_traits::ToPrimitive;
 use thiserror::Error;
+
+pub use cow::{CowArrayD, JArrayCow};
 
 // TODO: https://code.jsoftware.com/wiki/Vocabulary/ErrorMessages
 #[derive(Debug, Error)]
@@ -94,8 +98,6 @@ impl JError {
     }
 }
 
-type CowArrayD<'t, T> = CowArray<'t, T, IxDyn>;
-
 // All terminology should match J terminology:
 // Glossary: https://code.jsoftware.com/wiki/Vocabulary/Glossary
 // A Word is a part of speech.
@@ -125,18 +127,6 @@ pub enum JArray {
     FloatArray(ArrayD<f64>),
     ComplexArray(ArrayD<Complex64>),
     BoxArray(ArrayD<Word>),
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum JArrayCow<'a> {
-    BoolArray(CowArrayD<'a, u8>),
-    CharArray(CowArrayD<'a, char>),
-    IntArray(CowArrayD<'a, i64>),
-    ExtIntArray(CowArrayD<'a, BigInt>),
-    RationalArray(CowArrayD<'a, BigRational>),
-    FloatArray(CowArrayD<'a, f64>),
-    ComplexArray(CowArrayD<'a, Complex64>),
-    BoxArray(CowArrayD<'a, Word>),
 }
 
 impl JArray {
@@ -316,44 +306,6 @@ macro_rules! homo_array {
             })
             .collect::<Result<Vec<_>>>()?
     };
-}
-
-impl<'v> JArrayCow<'v> {
-    pub fn len(&self) -> usize {
-        match self {
-            JArrayCow::IntArray(x) => x.len(),
-            _ => todo!(),
-        }
-    }
-
-    pub fn shape(&self) -> &[usize] {
-        match self {
-            JArrayCow::IntArray(x) => x.shape(),
-            _ => todo!(),
-        }
-    }
-
-    pub fn outer_iter(&self) -> impl Iterator<Item = JArrayCow> + Clone {
-        match self {
-            JArrayCow::IntArray(x) => x.outer_iter().map(|x| x.into()),
-            _ => todo!(),
-        }
-    }
-}
-
-impl<'v> From<JArrayCow<'v>> for JArray {
-    fn from(value: JArrayCow<'v>) -> Self {
-        match value {
-            JArrayCow::IntArray(v) => JArray::IntArray(v.into_owned()),
-            _ => todo!(),
-        }
-    }
-}
-
-impl<'v> From<ArrayViewD<'v, i64>> for JArrayCow<'v> {
-    fn from(value: ArrayViewD<'v, i64>) -> Self {
-        JArrayCow::IntArray(value.into())
-    }
 }
 
 impl JArray {
